Move XdgOutputManagerV1 declaration from the header to the source, and provide a builder function
